test: #102-ai-note-indexing 인덱싱 트랜잭션 분리 및 실패 기록 검증 추가
