+++ src/test/java/com/keepgoing/keepgoing/ai/service/AiNoteIndexingServiceTest.java
@@ -1,30 +1,17 @@
 package com.keepgoing.keepgoing.ai.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
 import com.keepgoing.keepgoing.activity.service.ActivityDashboardService;
-import com.keepgoing.keepgoing.ai.domain.AiNoteChunk;
-import com.keepgoing.keepgoing.ai.domain.AiNoteIndex;
-import com.keepgoing.keepgoing.ai.domain.AiNoteIndexStatus;
-import com.keepgoing.keepgoing.ai.repository.AiNoteChunkRepository;
-import com.keepgoing.keepgoing.ai.repository.AiNoteIndexRepository;
-import com.keepgoing.keepgoing.ai.repository.AiNoteIndexingSourceView;
-import com.keepgoing.keepgoing.note.repository.NoteRepository;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
@@ -34,16 +21,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class AiNoteIndexingServiceTest {
 
 	@Mock
-	NoteRepository noteRepository;
+	AiNoteIndexingProcessor aiNoteIndexingProcessor;
 
 	@Mock
-	AiNoteIndexRepository aiNoteIndexRepository;
-
-	@Mock
-	AiNoteChunkRepository aiNoteChunkRepository;
-
-	@Spy
-	AiNoteChunker aiNoteChunker = new AiNoteChunker();
+	AiNoteIndexFailureRecorder aiNoteIndexFailureRecorder;
 
 	@Spy
 	Clock clock = Clock.fixed(
@@ -54,203 +35,44 @@ class AiNoteIndexingServiceTest {
 	@InjectMocks
 	AiNoteIndexingService aiNoteIndexingService;
 
-	@Nested
-	@DisplayName("인덱싱 처리")
-	class Process {
+	@Test
+	@DisplayName("인덱싱 처리가 성공하면 실패 상태를 기록하지 않는다")
+	void doesNotRecordFailureWhenProcessorSucceeds() {
+		// given
+		Long noteId = 10L;
 
-		@Test
-		@DisplayName("수집 가능한 노트면 chunk를 저장하고 COMPLETED로 마킹한다")
-		void marksCompletedWhenCollectable() {
-			// given
-			Long noteId = 10L;
-			Long authorId = 1L;
-			LocalDateTime requestedAt = LocalDateTime.of(2026, 5, 14, 9, 0);
+		// when
+		aiNoteIndexingService.process(noteId);
 
-			AiNoteIndex index = AiNoteIndex.pending(noteId, authorId, requestedAt);
-
-			given(aiNoteIndexRepository.findById(noteId)).willReturn(Optional.of(index));
-			given(noteRepository.findAiIndexingSourceById(noteId))
-					.willReturn(Optional.of(source(
-							noteId,
-							authorId,
-							"회의록",
-							"오늘 논의한 내용이 길게 들어간다. ".repeat(30),
-							true,
-							null,
-							requestedAt.plusHours(1)
-					)));
-			given(aiNoteChunkRepository.saveAll(any())).willAnswer(invocation -> invocation.getArgument(0));
-
-			// when
-			aiNoteIndexingService.process(noteId);
-
-			// then
-			verify(aiNoteChunkRepository).deleteByNoteId(noteId);
-
-			ArgumentCaptor<List<AiNoteChunk>> chunkCaptor = ArgumentCaptor.forClass(List.class);
-			verify(aiNoteChunkRepository).saveAll(chunkCaptor.capture());
-
-			List<AiNoteChunk> savedChunks = chunkCaptor.getValue();
-			assertThat(savedChunks).isNotEmpty();
-			assertThat(savedChunks.get(0).getNoteId()).isEqualTo(noteId);
-			assertThat(savedChunks.get(0).getAuthorId()).isEqualTo(authorId);
-			assertThat(savedChunks.get(0).getTitle()).isEqualTo("회의록");
-
-			assertThat(index.getStatus()).isEqualTo(AiNoteIndexStatus.COMPLETED);
-			assertThat(index.getIndexedAt()).isNotNull();
-			assertThat(index.getLastError()).isNull();
-			assertThat(index.getAttemptCount()).isEqualTo(1);
-		}
-
-		@Test
-		@DisplayName("삭제된 노트면 chunk를 제거하고 REMOVED로 마킹한다")
-		void marksRemovedWhenDeleted() {
-			// given
-			Long noteId = 10L;
-			Long authorId = 1L;
-			LocalDateTime requestedAt = LocalDateTime.of(2026, 5, 14, 9, 0);
-
-			AiNoteIndex index = AiNoteIndex.pending(noteId, authorId, requestedAt);
-
-			given(aiNoteIndexRepository.findById(noteId)).willReturn(Optional.of(index));
-			given(noteRepository.findAiIndexingSourceById(noteId))
-					.willReturn(Optional.of(source(
-							noteId,
-							authorId,
-							"삭제된 노트",
-							"내용",
-							true,
-							LocalDateTime.of(2026, 5, 14, 8, 0),
-							requestedAt.plusHours(1)
-					)));
-
-			// when
-			aiNoteIndexingService.process(noteId);
-
-			// then
-			verify(aiNoteChunkRepository).deleteByNoteId(noteId);
-			verify(aiNoteChunkRepository, never()).saveAll(any());
-
-			assertThat(index.getStatus()).isEqualTo(AiNoteIndexStatus.REMOVED);
-			assertThat(index.getIndexedAt()).isNull();
-			assertThat(index.getAttemptCount()).isEqualTo(1);
-		}
-
-		@Test
-		@DisplayName("AI 수집 대상이 아니면 chunk를 제거하고 REMOVED로 마킹한다")
-		void marksRemovedWhenNotCollectable() {
-			// given
-			Long noteId = 10L;
-			Long authorId = 1L;
-			LocalDateTime requestedAt = LocalDateTime.of(2026, 5, 14, 9, 0);
-
-			AiNoteIndex index = AiNoteIndex.pending(noteId, authorId, requestedAt);
-
-			given(aiNoteIndexRepository.findById(noteId)).willReturn(Optional.of(index));
-			given(noteRepository.findAiIndexingSourceById(noteId))
-					.willReturn(Optional.of(source(
-							noteId,
-							authorId,
-							"비수집 노트",
-							"내용",
-							false,
-							null,
-							requestedAt.plusHours(1)
-					)));
-
-			// when
-			aiNoteIndexingService.process(noteId);
-
-			// then
-			verify(aiNoteChunkRepository).deleteByNoteId(noteId);
-			verify(aiNoteChunkRepository, never()).saveAll(any());
-
-			assertThat(index.getStatus()).isEqualTo(AiNoteIndexStatus.REMOVED);
-			assertThat(index.getAttemptCount()).isEqualTo(1);
-		}
-
-		@Test
-		@DisplayName("인덱싱 중 예외가 발생하면 FAILED로 마킹한다")
-		void marksFailedWhenExceptionOccurs() {
-			// given
-			Long noteId = 10L;
-			Long authorId = 1L;
-			LocalDateTime requestedAt = LocalDateTime.of(2026, 5, 14, 9, 0);
-
-			AiNoteIndex index = AiNoteIndex.pending(noteId, authorId, requestedAt);
-
-			given(aiNoteIndexRepository.findById(noteId)).willReturn(Optional.of(index));
-			given(noteRepository.findAiIndexingSourceById(noteId))
-					.willThrow(new RuntimeException("db read fail"));
-
-			// when & then
-			assertThatThrownBy(() -> aiNoteIndexingService.process(noteId))
-					.isInstanceOf(RuntimeException.class)
-					.hasMessageContaining("db read fail");
-
-			assertThat(index.getStatus()).isEqualTo(AiNoteIndexStatus.FAILED);
-			assertThat(index.getLastError()).contains("db read fail");
-			assertThat(index.getAttemptCount()).isEqualTo(1);
-		}
-
-		@Test
-		@DisplayName("인덱스가 없으면 예외가 발생한다")
-		void throwsWhenIndexNotFound() {
-			// given
-			Long noteId = 10L;
-			given(aiNoteIndexRepository.findById(noteId)).willReturn(Optional.empty());
-
-			// when & then
-			assertThatThrownBy(() -> aiNoteIndexingService.process(noteId))
-					.isInstanceOf(IllegalStateException.class)
-					.hasMessageContaining("AI note index not found");
-		}
+		// then
+		verify(aiNoteIndexingProcessor).processInTransaction(noteId);
+		verify(aiNoteIndexFailureRecorder, never())
+				.markFailed(
+						org.mockito.ArgumentMatchers.any(),
+						org.mockito.ArgumentMatchers.any(),
+						org.mockito.ArgumentMatchers.any()
+				);
 	}
 
-	private AiNoteIndexingSourceView source(
-			Long noteId,
-			Long authorId,
-			String title,
-			String content,
-			boolean aiCollectable,
-			LocalDateTime deletedAt,
-			LocalDateTime updatedAt
-	) {
-		return new AiNoteIndexingSourceView() {
-			@Override
-			public Long getNoteId() {
-				return noteId;
-			}
+	@Test
+	@DisplayName("인덱싱 처리 중 예외가 발생하면 FAILED 상태를 기록하고 예외를 다시 던진다")
+	void recordsFailureAndRethrowsWhenProcessorFails() {
+		// given
+		Long noteId = 10L;
+		RuntimeException exception = new RuntimeException("db read fail");
 
-			@Override
-			public Long getAuthorId() {
-				return authorId;
-			}
+		willThrow(exception)
+				.given(aiNoteIndexingProcessor)
+				.processInTransaction(noteId);
 
-			@Override
-			public String getTitle() {
-				return title;
-			}
+		// when & then
+		assertThatThrownBy(() -> aiNoteIndexingService.process(noteId))
+				.isSameAs(exception);
 
-			@Override
-			public String getContent() {
-				return content;
-			}
-
-			@Override
-			public boolean isAiCollectable() {
-				return aiCollectable;
-			}
-
-			@Override
-			public LocalDateTime getDeletedAt() {
-				return deletedAt;
-			}
-
-			@Override
-			public LocalDateTime getUpdatedAt() {
-				return updatedAt;
-			}
-		};
+		verify(aiNoteIndexFailureRecorder).markFailed(
+				noteId,
+				LocalDateTime.of(2026, 5, 14, 9, 0),
+				"db read fail"
+		);
 	}
 }
